Fixed int overflow in C++ histogram buffer sizes and offsets
`omp_get_max_threads() * n_slices` was multiplied in `int` before being
widened to `size_t`, so e.g. 64 threads x 40M slices overflowed and the
per-thread histogram buffer was allocated too small. Same for the row
offsets, and for `2 * n_poles` indices in poles.cpp. No test: triggering
it needs an allocation of more than 2^31 elements. Found by clang-tidy
(bugprone-implicit-widening-of-multiplication-result).

+++ blond/core/backends/cpp/histogram.cpp
@@ -28,10 +28,10 @@ extern "C" void histogram(const real_t *__restrict__ input,
   // single bin can collect more than 2^31 - 1 particles
   index_t **histo =
       (index_t **)malloc(omp_get_max_threads() * sizeof(index_t *));
-  histo[0] =
-      (index_t *)malloc(omp_get_max_threads() * n_slices * sizeof(index_t));
+  histo[0] = (index_t *)malloc((size_t)omp_get_max_threads() * n_slices *
+                               sizeof(index_t));
   for (int i = 0; i < omp_get_max_threads(); i++)
-    histo[i] = (*histo + n_slices * i);
+    histo[i] = (*histo + (size_t)n_slices * i);
 
 #pragma omp parallel
   {
@@ -95,10 +95,10 @@ extern "C" void smooth_histogram(const real_t *__restrict__ input,
 
   // memory alloc for per thread histo
   real_t **histo = (real_t **)malloc(omp_get_max_threads() * sizeof(real_t *));
-  histo[0] =
-      (real_t *)malloc(omp_get_max_threads() * n_slices * sizeof(real_t));
+  histo[0] = (real_t *)malloc((size_t)omp_get_max_threads() * n_slices *
+                              sizeof(real_t));
   for (int i = 0; i < omp_get_max_threads(); i++)
-    histo[i] = (*histo + n_slices * i);
+    histo[i] = (*histo + (size_t)n_slices * i);
 
 #pragma omp parallel
   {
